net: stmmac: Deprecate snps,phy-addr and default to PHY auto-detection
The snps,phy-addr device tree property is non-standard, and should be
removed in favor of proper phy node support. Remove it from the binding
documents and warn if the property is still used.

Most PHYs respond to address 0, but a few don't, so default to PHY
address auto-detection, to make up for the lack of explicit address
selection.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>