touchscreen devicetree binding: Add touchscreen-swapped-x-y property
On devices with a native portrait screen a landscape touchscreen / digitizer
may be used, this happens e.g. on ebook readers. In this case the X and Y
axis of the touchscreen are swapped compared to the screen.

Add a touchscreen-swapped-x-y property which drivers can use to see if they
need to swap the axis to make the touchscreen coordinates match the screen
coordinates.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Changes in v3:
-s/touchscreen-swap-x-y/touchscreen-swapped-x-y/