phy-sun4i-usb: Activate pullups on init and add set vbus / id detect helpers
The usb0 phy is an otg phy, which allows explicit control over the pullups,
enable these on phy_init and disable them on phy_exit.

Besides the pullups the usb0 phy also has some status bits which are used to
signal the status of the id and vbus detect pins to the musb-core, add 2
sun4i-usb-phy specific helper functions for setting these bits.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>