ARM: dts: sun7i: Enable USB DRC on Bananapi (BROKEN!)
Enable the otg/drc usb controller on the Bananapi.

This is based on the fex file settings which claim that vbus-det is
PH5, but this does not work and the schematic shows that it is actually
connected to gpio1 of the axp209, so we cannot add otg support for the
bananapi until we get support for the axp209 gpio pins in the kernel.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>